+++ src/app/App.tsx
@@ -1,11 +1,11 @@
 import PostsManagerPage from '@/pages/PostsManagerPage.tsx';
 import Footer from '@/widgets/ui/Footer.tsx';
 import Header from '@/widgets/ui/Header.tsx';
-import { BrowserRouter as Router } from 'react-router-dom';
+import { AppRouter } from './routers';
 
 const App = () => {
   return (
-    <Router>
+    <AppRouter>
       <div className="flex flex-col min-h-screen">
         <Header />
         <main className="flex-grow container mx-auto px-4 py-8">
@@ -13,7 +13,7 @@ const App = () => {
         </main>
         <Footer />
       </div>
-    </Router>
+    </AppRouter>
   );
 };
