+++ src/app/routers/index.ts
@@ -1,0 +1,1 @@
+export { AppRouter } from './AppRouter';
