+++ src/index.tsx
@@ -1,4 +1,4 @@
-import App from '@/App';
+import App from '@/app/App';
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter as Router } from 'react-router-dom';
