+++ src/main.tsx
@@ -1,4 +1,4 @@
-import App from '@/App.tsx';
+import App from '@/app/App';
 import '@/index.css';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
